switch to double

+++ imageProcess.cpp
@@ -9,18 +9,18 @@ CImageProcess::~CImageProcess()
 {
 }
 
-float* CImageProcess::createRGBImage(int resX,int resY)
+double* CImageProcess::createRGBImage(int resX,int resY)
 {
-    float* retVal=new float[resX*resY*3];
+    double* retVal=new double[resX*resY*3];
     return(retVal);
 }
-void CImageProcess::copyRGBImage(int resX,int resY,float* imageSource,float* imageDest)
+void CImageProcess::copyRGBImage(int resX,int resY,double* imageSource,double* imageDest)
 {
     int v=resX*resY*3;
     for (int i=0;i<v;i++)
         imageDest[i]=imageSource[i];
 }
-void CImageProcess::clearRGBImage(int resX,int resY,float* image,float clearRed,float clearGreen,float clearBlue)
+void CImageProcess::clearRGBImage(int resX,int resY,double* image,double clearRed,double clearGreen,double clearBlue)
 {
     int v=resX*resY;
     for (int i=0;i<v;i++)
@@ -30,34 +30,34 @@ void CImageProcess::clearRGBImage(int resX,int resY,float* image,float clearRed,
         image[3*i+2]=clearBlue;
     }
 }
-float* CImageProcess::createIntensityImage(int resX,int resY)
+double* CImageProcess::createIntensityImage(int resX,int resY)
 {
-    float* retVal=new float[resX*resY];
+    double* retVal=new double[resX*resY];
     return(retVal);
 }
-void CImageProcess::copyIntensityImage(int resX,int resY,float* intensImageSource,float* intensImageDest)
+void CImageProcess::copyIntensityImage(int resX,int resY,double* intensImageSource,double* intensImageDest)
 {
     int v=resX*resY;
     for (int i=0;i<v;i++)
         intensImageDest[i]=intensImageSource[i];
 }
-void CImageProcess::clearIntensityImage(int resX,int resY,float* intensImage,float clearIntens)
+void CImageProcess::clearIntensityImage(int resX,int resY,double* intensImage,double clearIntens)
 {
     int v=resX*resY;
     for (int i=0;i<v;i++)
         intensImage[i]=clearIntens;
 }
-void CImageProcess::deleteImage(float* image)
+void CImageProcess::deleteImage(double* image)
 {
     delete[] image;
 }
-void CImageProcess::rgbImageToIntensityImage(int resX,int resY,float* rgbImage,float* intensImage)
+void CImageProcess::rgbImageToIntensityImage(int resX,int resY,double* rgbImage,double* intensImage)
 {
     int v=resX*resY;
     for (int i=0;i<v;i++)
         intensImage[i]=(rgbImage[3*i+0]+rgbImage[3*i+1]+rgbImage[3*i+2])/3.0f;
 }
-void CImageProcess::intensityImageToRGBImage(int resX,int resY,float* rgbImage,float* intensImage)
+void CImageProcess::intensityImageToRGBImage(int resX,int resY,double* rgbImage,double* intensImage)
 {
     int v=resX*resY;
     for (int i=0;i<v;i++)
@@ -68,15 +68,15 @@ void CImageProcess::intensityImageToRGBImage(int resX,int resY,float* rgbImage,f
     }
 }
 
-void CImageProcess::filter3x3RgbImage(int resX,int resY,float* rgbIn,float* rgbOut,float m[9])
+void CImageProcess::filter3x3RgbImage(int resX,int resY,double* rgbIn,double* rgbOut,double m[9])
 { // m is specified line after line
-    float w,cnt;
+    double w,cnt;
     int x,y,tmp;
     for (int i=0;i<resX;i++)
     {
         for (int j=0;j<resY;j++)
         {
-            float rgbCumul[3]={0.0f,0.0f,0.0f};
+            double rgbCumul[3]={0.0f,0.0f,0.0f};
             cnt=0.0f;
             for (int k=-1;k<2;k++)
             {
@@ -107,15 +107,15 @@ void CImageProcess::filter3x3RgbImage(int resX,int resY,float* rgbIn,float* rgbO
     }
 }
 
-void CImageProcess::filter5x5RgbImage(int resX,int resY,float* rgbIn,float* rgbOut,float m[25])
+void CImageProcess::filter5x5RgbImage(int resX,int resY,double* rgbIn,double* rgbOut,double m[25])
 { // m is specified line after line
-    float w,cnt;
+    double w,cnt;
     int x,y,tmp;
     for (int i=0;i<resX;i++)
     {
         for (int j=0;j<resY;j++)
         {
-            float rgbCumul[3]={0.0f,0.0f,0.0f};
+            double rgbCumul[3]={0.0f,0.0f,0.0f};
             cnt=0.0f;
             for (int k=-2;k<3;k++)
             {
@@ -146,7 +146,7 @@ void CImageProcess::filter5x5RgbImage(int resX,int resY,float* rgbIn,float* rgbO
     }
 }
 
-void CImageProcess::clampRgbImage(int resX,int resY,float* rgbImage,float lowClamp,float highClamp)
+void CImageProcess::clampRgbImage(int resX,int resY,double* rgbImage,double lowClamp,double highClamp)
 {
     int s=resX*resY*3;
     for (int i=0;i<s;i++)
@@ -161,15 +161,15 @@ void CImageProcess::clampRgbImage(int resX,int resY,float* rgbImage,float lowCla
     }
 }
 
-void CImageProcess::getEdges(int resX,int resY,float* imageSource,float* imageDest)
+void CImageProcess::getEdges(int resX,int resY,double* imageSource,double* imageDest)
 {
-    float maxV=0.0f;
+    double maxV=0.0f;
     for (int i=0;i<resX;i++)
     {
         for (int j=0;j<resY;j++)
         {
-            float gx=0.0f;
-            float gy=0.0f;
+            double gx=0.0f;
+            double gy=0.0f;
             int gxc=0;
             int gyc=0;
 
@@ -215,7 +215,7 @@ void CImageProcess::getEdges(int resX,int resY,float* imageSource,float* imageDe
                 gx=0.0f;
             if (gyc==0)
                 gy=0.0f;
-            float g=sqrtf(gx*gx+gy*gy);
+            double g=sqrtf(gx*gx+gy*gy);
             imageDest[i+j*resX]=g;
             if (g>maxV)
                 maxV=g;
@@ -225,7 +225,7 @@ void CImageProcess::getEdges(int resX,int resY,float* imageSource,float* imageDe
         scaleIntensity(resX,resY,imageDest,1.0f/maxV,false);
 }
 
-void CImageProcess::scaleIntensity(int resX,int resY,float* image,float scaleFactor,bool clampToMax)
+void CImageProcess::scaleIntensity(int resX,int resY,double* image,double scaleFactor,bool clampToMax)
 {
     int v=resX*resY;
     if (clampToMax)
@@ -244,7 +244,7 @@ void CImageProcess::scaleIntensity(int resX,int resY,float* image,float scaleFac
     }
 }
 
-void CImageProcess::scaleRGB(int resX,int resY,float* image,float redFactor,float greenFactor,float blueFactor,bool clampToMax)
+void CImageProcess::scaleRGB(int resX,int resY,double* image,double redFactor,double greenFactor,double blueFactor,bool clampToMax)
 {
     int v=resX*resY;
     if (clampToMax)
@@ -273,25 +273,25 @@ void CImageProcess::scaleRGB(int resX,int resY,float* image,float redFactor,floa
     }
 }
 
-void CImageProcess::scaleRgbImageWithIntensityImage(int resX,int resY,float* rgbImage,float* intensImage)
+void CImageProcess::scaleRgbImageWithIntensityImage(int resX,int resY,double* rgbImage,double* intensImage)
 {
     int s=resX*resY;
     for (int i=0;i<s;i++)
     {
-        float d=intensImage[i]*3.0f;
+        double d=intensImage[i]*3.0f;
         rgbImage[3*i+0]*=d;
         rgbImage[3*i+1]*=d;
         rgbImage[3*i+2]*=d;
     }
 }
 
-void CImageProcess::boxBlurIntensity(int resX,int resY,float* imageSource,float* imageDest)
+void CImageProcess::boxBlurIntensity(int resX,int resY,double* imageSource,double* imageDest)
 {
     for (int i=0;i<resX;i++)
     {
         for (int j=0;j<resY;j++)
         {
-            float gx=0.0f;
+            double gx=0.0f;
             int gxc=0;
             if (i>0)
             {
@@ -336,21 +336,21 @@ void CImageProcess::boxBlurIntensity(int resX,int resY,float* imageSource,float*
                     gxc++;
                 }
             }
-            gx/=float(gxc);
+            gx/=double(gxc);
             imageDest[i+j*resX]=gx;
         }
     }
 }
 
-void CImageProcess::boxBlurRGB(int resX,int resY,float* imageSource,float* imageDest)
+void CImageProcess::boxBlurRGB(int resX,int resY,double* imageSource,double* imageDest)
 {
     for (int i=0;i<resX;i++)
     {
         for (int j=0;j<resY;j++)
         {
-            float gRed=0.0f;
-            float gGreen=0.0f;
-            float gBlue=0.0f;
+            double gRed=0.0f;
+            double gGreen=0.0f;
+            double gBlue=0.0f;
             int gxc=0;
             if (i>0)
             {
@@ -413,9 +413,9 @@ void CImageProcess::boxBlurRGB(int resX,int resY,float* imageSource,float* image
                     gxc++;
                 }
             }
-            gRed/=float(gxc);
-            gGreen/=float(gxc);
-            gBlue/=float(gxc);
+            gRed/=double(gxc);
+            gGreen/=double(gxc);
+            gBlue/=double(gxc);
             imageDest[3*(i+j*resX)+0]=gRed;
             imageDest[3*(i+j*resX)+1]=gGreen;
             imageDest[3*(i+j*resX)+2]=gBlue;
@@ -423,9 +423,9 @@ void CImageProcess::boxBlurRGB(int resX,int resY,float* imageSource,float* image
     }
 }
 
-void CImageProcess::blurIntensity(int resX,int resY,float* image)
+void CImageProcess::blurIntensity(int resX,int resY,double* image)
 {
-    float* im2=createIntensityImage(resX,resY);
+    double* im2=createIntensityImage(resX,resY);
     copyIntensityImage(resX,resY,image,im2);
     boxBlurIntensity(resX,resY,im2,image);
     boxBlurIntensity(resX,resY,image,im2);
@@ -433,9 +433,9 @@ void CImageProcess::blurIntensity(int resX,int resY,float* image)
     deleteImage(im2);
 }
 
-void CImageProcess::blurRGB(int resX,int resY,float* image)
+void CImageProcess::blurRGB(int resX,int resY,double* image)
 {
-    float* im2=createRGBImage(resX,resY);
+    double* im2=createRGBImage(resX,resY);
     copyRGBImage(resX,resY,image,im2);
     boxBlurRGB(resX,resY,im2,image);
     boxBlurRGB(resX,resY,image,im2);
@@ -443,9 +443,9 @@ void CImageProcess::blurRGB(int resX,int resY,float* image)
     deleteImage(im2);
 }
 
-void CImageProcess::horizontalFlipRGB(int resX,int resY,float* image)
+void CImageProcess::horizontalFlipRGB(int resX,int resY,double* image)
 {
-    float tmp;
+    double tmp;
     for (int i=0;i<resX/2;i++)
     {
         for (int j=0;j<resY;j++)
@@ -460,9 +460,9 @@ void CImageProcess::horizontalFlipRGB(int resX,int resY,float* image)
     }
 }
 
-void CImageProcess::verticalFlipRGB(int resX,int resY,float* image)
+void CImageProcess::verticalFlipRGB(int resX,int resY,double* image)
 {
-    float tmp;
+    double tmp;
     for (int i=0;i<resX;i++)
     {
         for (int j=0;j<resY/2;j++)
@@ -477,7 +477,7 @@ void CImageProcess::verticalFlipRGB(int resX,int resY,float* image)
     }
 }
 
-void CImageProcess::keepThresholdIntensity(int resX,int resY,float* image,float thresholdVal,bool keepAbove)
+void CImageProcess::keepThresholdIntensity(int resX,int resY,double* image,double thresholdVal,bool keepAbove)
 {
     int v=resX*resY;
     if (keepAbove)
@@ -498,7 +498,7 @@ void CImageProcess::keepThresholdIntensity(int resX,int resY,float* image,float 
     }
 }
 
-void CImageProcess::keepThresholdRGB(int resX,int resY,float* image,float thresholdVal,bool keepAbove)
+void CImageProcess::keepThresholdRGB(int resX,int resY,double* image,double thresholdVal,bool keepAbove)
 {
     int v=resX*resY;
     thresholdVal*=3.0f;
@@ -506,7 +506,7 @@ void CImageProcess::keepThresholdRGB(int resX,int resY,float* image,float thresh
     {
         for (int i=0;i<v;i++)
         {
-            float w=image[3*i+0]+image[3*i+1]+image[3*i+2];
+            double w=image[3*i+0]+image[3*i+1]+image[3*i+2];
             if (w<thresholdVal)
             {
                 image[3*i+0]=0.0f;
@@ -519,7 +519,7 @@ void CImageProcess::keepThresholdRGB(int resX,int resY,float* image,float thresh
     {
         for (int i=0;i<v;i++)
         {
-            float w=image[3*i+0]+image[3*i+1]+image[3*i+2];
+            double w=image[3*i+0]+image[3*i+1]+image[3*i+2];
             if (w>thresholdVal)
             {
                 image[3*i+0]=0.0f;
@@ -530,7 +530,7 @@ void CImageProcess::keepThresholdRGB(int resX,int resY,float* image,float thresh
     }
 }
 
-void CImageProcess::nonZeroToOneIntensity(int resX,int resY,float* image)
+void CImageProcess::nonZeroToOneIntensity(int resX,int resY,double* image)
 {
     int v=resX*resY;
     for (int i=0;i<v;i++)
@@ -540,12 +540,12 @@ void CImageProcess::nonZeroToOneIntensity(int resX,int resY,float* image)
     }
 }
 
-void CImageProcess::nonZeroToOneRGB(int resX,int resY,float* image)
+void CImageProcess::nonZeroToOneRGB(int resX,int resY,double* image)
 {
     int v=resX*resY;
     for (int i=0;i<v;i++)
     {
-        float w=image[3*i+0]+image[3*i+1]+image[3*i+2];
+        double w=image[3*i+0]+image[3*i+1]+image[3*i+2];
         if (w!=0.0f)
         {
             image[3*i+0]=1.0f;
@@ -555,14 +555,14 @@ void CImageProcess::nonZeroToOneRGB(int resX,int resY,float* image)
     }
 }
 
-void CImageProcess::addImagesRGB(int resX,int resY,float* imageInOut,float* imageOverlay)
+void CImageProcess::addImagesRGB(int resX,int resY,double* imageInOut,double* imageOverlay)
 {
     int v=resX*resY*3;
     for (int i=0;i<v;i++)
         imageInOut[i]+=imageOverlay[i];
 }
 
-void CImageProcess::clampToOneRGB(int resX,int resY,float* image)
+void CImageProcess::clampToOneRGB(int resX,int resY,double* image)
 {
     int v=resX*resY*3;
     for (int i=0;i<v;i++)
@@ -572,23 +572,23 @@ void CImageProcess::clampToOneRGB(int resX,int resY,float* image)
     }
 }
 
-void CImageProcess::predef_lightBlurRGB(int resX,int resY,float* image)
+void CImageProcess::predef_lightBlurRGB(int resX,int resY,double* image)
 {
     blurRGB(resX,resY,image);
 }
 
-void CImageProcess::predef_heavyBlurRGB(int resX,int resY,float* image)
+void CImageProcess::predef_heavyBlurRGB(int resX,int resY,double* image)
 {
     blurRGB(resX,resY,image);
     blurRGB(resX,resY,image);
     blurRGB(resX,resY,image);
 }
 
-void CImageProcess::predef_getThinEdgesRGB(int resX,int resY,float* image)
+void CImageProcess::predef_getThinEdgesRGB(int resX,int resY,double* image)
 {
-    float* im0=createIntensityImage(resX,resY);
+    double* im0=createIntensityImage(resX,resY);
     rgbImageToIntensityImage(resX,resY,image,im0);
-    float* im1=createIntensityImage(resX,resY);
+    double* im1=createIntensityImage(resX,resY);
     getEdges(resX,resY,im0,im1);
     keepThresholdIntensity(resX,resY,im1,0.1f,true);
     nonZeroToOneIntensity(resX,resY,im1);
@@ -597,15 +597,15 @@ void CImageProcess::predef_getThinEdgesRGB(int resX,int resY,float* image)
     deleteImage(im0);
 }
 
-void CImageProcess::predef_getThinEdgeOverlayRGB(int resX,int resY,float* image)
+void CImageProcess::predef_getThinEdgeOverlayRGB(int resX,int resY,double* image)
 {
-    float* im0=createIntensityImage(resX,resY);
+    double* im0=createIntensityImage(resX,resY);
     rgbImageToIntensityImage(resX,resY,image,im0);
-    float* im1=createIntensityImage(resX,resY);
+    double* im1=createIntensityImage(resX,resY);
     getEdges(resX,resY,im0,im1);
     keepThresholdIntensity(resX,resY,im1,0.1f,true);
     nonZeroToOneIntensity(resX,resY,im1);
-    float* im2=createRGBImage(resX,resY);
+    double* im2=createRGBImage(resX,resY);
     copyRGBImage(resX,resY,image,im2);
     intensityImageToRGBImage(resX,resY,im2,im1);
     addImagesRGB(resX,resY,image,im2);
@@ -615,12 +615,12 @@ void CImageProcess::predef_getThinEdgeOverlayRGB(int resX,int resY,float* image)
     deleteImage(im0);
 }
 
-void CImageProcess::predef_getThickEdgesRGB(int resX,int resY,float* image)
+void CImageProcess::predef_getThickEdgesRGB(int resX,int resY,double* image)
 {
-    float* im0=createIntensityImage(resX,resY);
+    double* im0=createIntensityImage(resX,resY);
     rgbImageToIntensityImage(resX,resY,image,im0);
     blurIntensity(resX,resY,im0);
-    float* im1=createIntensityImage(resX,resY);
+    double* im1=createIntensityImage(resX,resY);
     getEdges(resX,resY,im0,im1);
     keepThresholdIntensity(resX,resY,im1,0.1f,true);
     nonZeroToOneIntensity(resX,resY,im1);
@@ -629,7 +629,7 @@ void CImageProcess::predef_getThickEdgesRGB(int resX,int resY,float* image)
     deleteImage(im0);
 }
 
-void CImageProcess::predef_invertRGB(int resX,int resY,float* image)
+void CImageProcess::predef_invertRGB(int resX,int resY,double* image)
 {
     int v=resX*resY*3;
     for (int i=0;i<v;i++)
